+++ core/src/include/filetypes.h
@@ -2,7 +2,7 @@
    BAREOS® - Backup Archiving REcovery Open Sourced
 
    Copyright (C) 2000-2010 Free Software Foundation Europe e.V.
-   Copyright (C) 2016-2021 Bareos GmbH & Co. KG
+   Copyright (C) 2016-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -28,6 +28,7 @@
 #ifndef BAREOS_INCLUDE_FILETYPES_H_
 #define BAREOS_INCLUDE_FILETYPES_H_
 
+#include <cstdint>
 
 /**
  *  File type (Bareos defined).
@@ -38,48 +39,55 @@
  *    used for the file type. The upper bits are used to indicate
  *    additional optional fields in the attribute record.
  */
-#define FT_MASK 0xFFFF  /**< Bits used by FT (type) */
-#define FT_LNKSAVED 1   /**< hard link to file already saved */
-#define FT_REGE 2       /**< Regular file but empty */
-#define FT_REG 3        /**< Regular file */
-#define FT_LNK 4        /**< Soft Link */
-#define FT_DIREND 5     /**< Directory at end (saved) */
-#define FT_SPEC 6       /**< Special file -- chr, blk, fifo, sock */
-#define FT_NOACCESS 7   /**< Not able to access */
-#define FT_NOFOLLOW 8   /**< Could not follow link */
-#define FT_NOSTAT 9     /**< Could not stat file */
-#define FT_NOCHG 10     /**< Incremental option, file not changed */
-#define FT_DIRNOCHG 11  /**< Incremental option, directory not changed */
-#define FT_ISARCH 12    /**< Trying to save archive file */
-#define FT_NORECURSE 13 /**< No recursion into directory */
-#define FT_NOFSCHG 14   /**< Different file system, prohibited */
-#define FT_NOOPEN 15    /**< Could not open directory */
-#define FT_RAW 16       /**< Raw block device */
-#define FT_FIFO 17      /**< Raw fifo device */
-/*
- * The DIRBEGIN packet is sent to the FD file processing routine so
- * that it can filter packets, but otherwise, it is not used
- * or saved */
-#define FT_DIRBEGIN 18      /**< Directory at beginning (not saved) */
-#define FT_INVALIDFS 19     /**< File system not allowed for */
-#define FT_INVALIDDT 20     /**< Drive type not allowed for */
-#define FT_REPARSE 21       /**< Win NTFS reparse point */
-#define FT_PLUGIN 22        /**< Plugin generated filename */
-#define FT_DELETED 23       /**< Deleted file entry */
-#define FT_BASE 24          /**< Duplicate base file entry */
-#define FT_RESTORE_FIRST 25 /**< Restore this "object" first */
-#define FT_JUNCTION 26      /**< Win32 Junction point */
-#define FT_PLUGIN_CONFIG 27 /**< Object for Plugin configuration */
-#define FT_PLUGIN_CONFIG_FILLED \
-  28 /**< Object for Plugin configuration filled by Director */
+constexpr int32_t FT_MASK = 0xFFFF; /**< Bits used by FT (type) */
+enum FILETYPES : int32_t
+{
+  FT_LNKSAVED = 1, /**< hard link to file already saved */
+  FT_REGE,         /**< Regular file but empty */
+  FT_REG,          /**< Regular file */
+  FT_LNK,          /**< Soft Link */
+  FT_DIREND,       /**< Directory at end (saved) */
+  FT_SPEC,         /**< Special file -- chr, blk, fifo, sock */
+  FT_NOACCESS,     /**< Not able to access */
+  FT_NOFOLLOW,     /**< Could not follow link */
+  FT_NOSTAT,       /**< Could not stat file */
+  FT_NOCHG,        /**< Incremental option, file not changed */
+  FT_DIRNOCHG,     /**< Incremental option, directory not changed */
+  FT_ISARCH,       /**< Trying to save archive file */
+  FT_NORECURSE,    /**< No recursion into directory */
+  FT_NOFSCHG,      /**< Different file system, prohibited */
+  FT_NOOPEN,       /**< Could not open directory */
+  FT_RAW,          /**< Raw block device */
+  FT_FIFO,         /**< Raw fifo device */
+  /*
+   * The DIRBEGIN packet is sent to the FD file processing routine so
+   * that it can filter packets, but otherwise, it is not used
+   * or saved */
+  FT_DIRBEGIN,             /**< Directory at beginning (not saved) */
+  FT_INVALIDFS,            /**< File system not allowed for */
+  FT_INVALIDDT,            /**< Drive type not allowed for */
+  FT_REPARSE,              /**< Win NTFS reparse point */
+  FT_PLUGIN,               /**< Plugin generated filename */
+  FT_DELETED,              /**< Deleted file entry */
+  FT_BASE,                 /**< Duplicate base file entry */
+  FT_RESTORE_FIRST,        /**< Restore this "object" first */
+  FT_JUNCTION,             /**< Win32 Junction point */
+  FT_PLUGIN_CONFIG,        /**< Object for Plugin configuration */
+  FT_PLUGIN_CONFIG_FILLED, /**< Object for Plugin configuration filled by
+                              Director */
+};
+static_assert(FT_PLUGIN_CONFIG_FILLED == 28);
 
-#define FT_UNSET FT_MASK /**< Initial value when not determined yet  */
+constexpr int32_t FT_UNSET
+    = FT_MASK; /**< Initial value when not determined yet  */
 /* Definitions for upper part of type word (see above). */
-#define AR_DATA_STREAM (1 << 16) /**< Data stream id present */
+constexpr int32_t AR_DATA_STREAM = (1 << 16); /**< Data stream id present */
 
 /* Quick way to know if a Filetype is about a plugin "Object" */
-#define IS_FT_OBJECT(x)                                          \
-  (((x) == FT_RESTORE_FIRST) || ((x) == FT_PLUGIN_CONFIG_FILLED) \
-   || ((x) == FT_PLUGIN_CONFIG))
+inline bool IS_FT_OBJECT(int32_t x)
+{
+  return x == FT_RESTORE_FIRST || x == FT_PLUGIN_CONFIG_FILLED
+         || x == FT_PLUGIN_CONFIG;
+}
 
 #endif  // BAREOS_INCLUDE_FILETYPES_H_