+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - github actions: PyPi: install setuptools [PR #1589]
 - restore: add fileregex parameter [PR #1587]
 - scripts: force cd / for all PostgreSQL scripts [PR #1607]
+- Improve python plugin configuration [PR #1605]
 
 ### Removed
 - plugins: remove old deprecated postgres plugin [PR #1606]
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR #1581]: https://github.com/bareos/bareos/pull/1581
 [PR #1587]: https://github.com/bareos/bareos/pull/1587
 [PR #1589]: https://github.com/bareos/bareos/pull/1589
+[PR #1605]: https://github.com/bareos/bareos/pull/1605
 [PR #1606]: https://github.com/bareos/bareos/pull/1606
 [PR #1607]: https://github.com/bareos/bareos/pull/1607
 [unreleased]: https://github.com/bareos/bareos/tree/master